tam_automation/requirements.txt
playwright>=1.40
gspread>=6.0
google-auth>=2.0
